requirements.txt
google-cloud-speech==0.28.0
grpcio==1.0.0
grpcio-tools
pyOpenSSL
cryptography
ws4py
blessings
requests
webrtcvad